script size.c

+++ hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 		long long int longlongtype;
 		char chartype;
 
-		printf("Size of int: %lu bytes\n", sizeof(inttype));
-		printf("Size of float: %lu bytes\n", sizeof(floattype));
-		printf("Size of long int: %lu bytes\n", sizeof(longtype));
-		printf("Size of char: %lu byte\n", sizeof(charType));
-		printf("Size of a long long int: %lu bytes\n", sizeof(longlongtype));
+		printf("Size of int: %lu bytes\n", (unsigned)sizeof(inttype));
+		printf("Size of float: %lu bytes\n", (unsigned)sizeof(floattype));
+		printf("Size of long int: %lu bytes\n", (unsigned)sizeof(longtype));
+		printf("Size of char: %lu byte\n", (unsigned)sizeof(charType));
+		printf("Size of a long long int: %lu bytes\n", (unsigned)sizeof(longlongtype));
 		return (0);
 }
